first extract from datalad PR

+++ nipoppy/workflows/dataset_init.py
@@ -4,6 +4,8 @@ import logging
 from pathlib import Path
 from typing import Optional
 
+import pandas as pd
+
 from nipoppy.env import LogColor, StrOrPathLike
 from nipoppy.utils import (
     DPATH_DESCRIPTORS,
@@ -24,6 +26,7 @@ class InitWorkflow(BaseWorkflow):
     def __init__(
         self,
         dpath_root: Path,
+        bids_source=None,
         fpath_layout: Optional[StrOrPathLike] = None,
         logger: Optional[logging.Logger] = None,
         dry_run: bool = False,
@@ -37,20 +40,49 @@ class InitWorkflow(BaseWorkflow):
             dry_run=dry_run,
         )
         self.fname_readme = "README.md"
+        self.bids_source = bids_source
 
     def run_main(self):
-        """Create dataset directory structure."""
+        """Create dataset directory structure.
+
+        Create directories and add a readme in each.
+        Copy boutiques descriptors and invocations.
+        Copy default config files.
+
+        If the BIDS source dataset is requested, it is installed with datalad.
+        """
         # dataset must not already exist
         if self.dpath_root.exists():
             raise FileExistsError("Dataset directory already exists")
 
         # create directories
         for dpath in self.layout.dpaths:
-            self.mkdir(dpath)
+
+            # If a bids_source is passed it means datalad is installed.
+            if self.bids_source is not None and dpath.stem == "bids":
+                self.logger.info(
+                    f"Installing datalad BIDS raw dataset from {self.bids_source}."
+                )
+                from datalad import api
+
+                dataset = None
+                if self.use_dalatad:
+                    dataset = self.dpath_root
+
+                api.install(
+                    dataset=dataset,
+                    path=dpath,
+                    source=self.bids_source,
+                    result_renderer="disabled",
+                )
+            else:
+                self.mkdir(dpath)
 
         for dpath, description in self.layout.dpath_descriptions:
             fpath_readme = dpath / self.fname_readme
-            if description is not None and not self.dry_run:
+            if (description is not None and not self.dry_run) and not (
+                self.bids_source is not None and dpath.stem == "bids"
+            ):
                 fpath_readme.write_text(f"{description}\n")
 
         # copy descriptor files
@@ -85,12 +117,64 @@ class InitWorkflow(BaseWorkflow):
             FPATH_SAMPLE_MANIFEST, self.layout.fpath_manifest, log_level=logging.DEBUG
         )
 
+        if self.bids_source is not None:
+            self._init_manifest_from_bids_dataset()
+
         # inform user to edit the sample files
         self.logger.warning(
             f"Sample config and manifest files copied to {self.layout.fpath_config}"
             f" and {self.layout.fpath_manifest} respectively. They should be edited"
             " to match your dataset"
         )
+
+    def _init_manifest_from_bids_dataset(self) -> None:
+        """Assume a BIDS dataset with session level folders.
+
+        No BIDS validation is done.
+        """
+        df = {"participant_id": [], "visit_id": [], "session_id": [], "datatype": []}
+        participant_ids = sorted(
+            [
+                x.name
+                for x in (self.layout.dpath_bids).iterdir()
+                if x.is_dir() and x.name.startswith("sub-")
+            ]
+        )
+
+        self.logger.info("Creating a manifest.csv from the BIDS dataset content.")
+
+        for ppt in participant_ids:
+
+            session_ids = sorted(
+                [
+                    x.name
+                    for x in (self.layout.dpath_bids / ppt).iterdir()
+                    if x.is_dir() and x.name.startswith("ses-")
+                ]
+            )
+            if not session_ids:
+                self.logger.warning(
+                    f"Skipping subject '{ppt}': could not find a session level folder."
+                )
+                continue
+
+            for ses in session_ids:
+                datatypes = sorted(
+                    [
+                        x.name
+                        for x in (self.layout.dpath_bids / ppt / ses).iterdir()
+                        if x.is_dir()
+                    ]
+                )
+
+                df["participant_id"].append(ppt.replace("sub-", ""))
+                df["session_id"].append(ses.replace("ses-", ""))
+                df["datatype"].append("[" + "'" + "' ,'".join(datatypes) + "'" + "]")
+
+        df["visit_id"] = df["session_id"]
+
+        df = pd.DataFrame(df)
+        df.to_csv(self.layout.fpath_manifest, index=False)
 
     def run_cleanup(self):
         """Log a success message."""
